Added support for pipelining
